add: authorization on routes

+++ src/app/api/users/[uuid]/route.tsx
@@ -1,27 +1,26 @@
-import { userUpdateSchema } from "@/types/user";
+import { HierarchyRole, userUpdateSchema } from "@/types/user";
 import { validateRequest } from "@/utils/server_api/requests/validate";
 import prisma from "@/utils/libs/prisma/single_instance";
-import type { NextRequest } from "next/server";
-import { jsonResponseNotFound, jsonResponseBadRequest } from "@/utils/server_api/responses/errors";
+import { jsonResponseNotFound, jsonResponseBadRequest, jsonResponseUnauthorized } from "@/utils/server_api/responses/errors";
 import { responseDelete, jsonResponsePatch } from "@/utils/server_api/responses/successes";
+import { getUserWithUuid } from "@/utils/libs/prisma/server_helper";
+import { isActionAuthorized } from "@/utils/helper";
+import { getCurrentUser } from "@/utils/libs/clerk/server_helper";
+import type { NextRequest } from "next/server";
 
 const ACCEPTED_CONTENT_TYPE = "application/json";
 
-export async function DELETE(_: NextRequest, { params }: { params: { uuid: string } }) {
+export async function DELETE(req: NextRequest, { params }: { params: { uuid: string } }) {
 	// we get the UUID from the URL params
 	const uuid = params.uuid;
 
 	if (!uuid) return jsonResponseNotFound("User not found");
 
-	const user = await prisma.user.findUnique({
-		where: {
-			uuid,
-		},
-	});
+	const user = await getUserWithUuid(uuid);
 
 	if (!user) return jsonResponseNotFound("User not found");
 
-	// TODO setup rights, only admin should be able to delete users
+	if (!isActionAuthorized(user)) return jsonResponseUnauthorized();	
 
 	await prisma.user.delete({
 		where: {
@@ -38,13 +37,11 @@ export async function PATCH(req: NextRequest, { params }: { params: { uuid: stri
 
 	if (!uuid) return jsonResponseNotFound("User not found");
 
-	const user = await prisma.user.findUnique({
-		where: {
-			uuid,
-		},
-	});
+	const user = await getUserWithUuid(uuid);
 
 	if (!user) return jsonResponseNotFound("User not found");
+
+	if (!isActionAuthorized(user)) return jsonResponseUnauthorized();	
 
 	const { error, messageError, data } = await validateRequest(
 		req,
@@ -53,8 +50,6 @@ export async function PATCH(req: NextRequest, { params }: { params: { uuid: stri
 	);
 
 	if (error) return jsonResponseBadRequest(messageError);
-
-	// TODO setup rights, only admin should be able to update users
 
 	const userUpdated = await prisma.user.update({
 		where: {
